Resign percent can be float; agent resign bugfix
Previously the -r <n> option was an integer, so it wasn't possible to
fix a value between 0 and 1%. Now sai accepts a flota value.

Bugfix: the previous commit on agent-aware resigning did not work when
lambda = mu and so in particular when lambda = mu = 0 (agent "zero").

+++ src/GTP.h
@@ -93,7 +93,7 @@ extern size_t cfg_max_tree_size;
 extern int cfg_max_cache_ratio_percent;
 extern TimeManagement::enabled_t cfg_timemanage;
 extern int cfg_lagbuffer_cs;
-extern int cfg_resignpct;
+extern float cfg_resignpct;
 extern float cfg_resign_threshold;
 extern int cfg_noise;
 extern bool cfg_fpuzero;
